Закончил функциональную часть

+++ pages/index.tsx
@@ -1,12 +1,13 @@
 import React, { useState } from 'react';
-import Head from 'next/head';
-import Layout from '../components/Layout';
 import styled from 'styled-components';
+import { GetServerSideProps } from 'next';
 import Router from 'next/router';
+import Head from 'next/head';
 
-import RangeSlider from '../components/Button';
+import RangeSlider from '../components/RangeSlider';
 import Counter from '../components/Counter';
-import Button from '../components/startButton';
+import Button from '../components/Button';
+import Layout from '../components/Layout';
 
 interface IIndexProps {
   words: string[];
@@ -34,9 +35,9 @@ const Table = styled.div`
   }
 `;
 
-export default function Index(props: IIndexProps) {
+export default function Index({ words }: IIndexProps) {
   const [state, setState] = useState({
-    wordsToShow: '5',
+    wordsCount: '5',
     distance: '25',
     letters: '5',
     distanceInc: '5',
@@ -51,12 +52,32 @@ export default function Index(props: IIndexProps) {
     setState({ ...state, speed: value });
   };
 
-  const gameStart = (e: React.MouseEvent<HTMLButtonElement>) => {
+  const wordSplit = (wordsArr) => {
+    const resultArr: string[] = [];
+
+    for (let i = 0; i < wordsArr.length; i++) {
+      const stringArr = wordsArr[i].split('');
+      stringArr.splice(Math.floor(stringArr.length / 2), 0, ';');
+      resultArr.push(stringArr.join(''));
+    }
+
+    return resultArr.map((v: string) => v.split(';'));
+  };
+
+  const gameStart = () => {
+    if (!words) {
+      return null;
+    }
     localStorage.clear();
-    const arr: string[] = props.words.filter((v) => v.length === +state.letters);
-    arr.sort(() => Math.random() - 0.5).splice(+state.wordsToShow);
-    localStorage.setItem('words', JSON.stringify(arr));
+
+    const arr: string[] = words.filter((v) => v.length === +state.letters);
+    arr.sort(() => Math.random() - 0.5).splice(+state.wordsCount);
+
+    const splitedWords = wordSplit(arr);
+
+    localStorage.setItem('words', JSON.stringify(splitedWords));
     localStorage.setItem('property', JSON.stringify(state));
+
     Router.push('/game');
   };
 
@@ -73,9 +94,9 @@ export default function Index(props: IIndexProps) {
               title="Сколько слов"
               min="1"
               max="10"
-              name="wordsToShow"
+              name="wordsCount"
               onChange={sliderChangeHandler}
-              value={state.wordsToShow}
+              value={state.wordsCount}
             />
             <RangeSlider
               title="Сколько букв в словах"
@@ -92,6 +113,7 @@ export default function Index(props: IIndexProps) {
               value={state.speed}
               onChange={speedChangeHandler}
             />
+            {!words && <p>Не установлено соединение с сервером</p>}
           </div>
           <div className="column">
             <RangeSlider
@@ -120,7 +142,14 @@ export default function Index(props: IIndexProps) {
   );
 }
 
-Index.getInitialProps = async () => {
+export const getServerSideProps: GetServerSideProps<IIndexProps> = async () => {
   const words = await fetch('http://localhost:4200/words').then((res) => res.json());
-  return { words };
+
+  if (Object.keys(words).length === 0) {
+    return {
+      props: { words: null },
+    };
+  }
+
+  return { props: { words } };
 };
